Do Not Attempt Decoding When We Know It Will Fail
This commit changes the `orDefault` and `orDefaultAccumulating` methods to not attempt decoding when we,

* Know we have a default value
* We know decoding will fail.

The code is made a bit more complex and less optimal because we have to run the decoder when the value is `null`, though we don't when it is missing. This is because the Circe `Decoder[Option[A]]` has a special case for this input `{"a": null}`. All other types would consider the value `null` and the missing key `"a"` to be the same case and use a default, but `Decoder[Option[A]]` considers `{"a": null}` as `Right(None)`.

See the test `"Option[T] with default should be None if null decoded"` for a complete example.

In the future, for defaulting codecs, we might consider treating a missing value and `JNull` the same in all cases, even for `Option`.

This commit also adds a benchmark package which was used to verify the issue referenced in https://github.com/circe/circe/issues/1972.

Note, this optimization works _around_ the issue in https://github.com/circe/circe/issues/1972, but there is still an issue in circe proper for https://github.com/circe/circe/issues/1972. This optimization is good with or without the issue described in https://github.com/circe/circe/issues/1972.